Extend fix for titles with linebreaks (BL-11717)
